fix file existence check

+++ manifests/init.pp
@@ -18,11 +18,7 @@ class drobo_dashboard {
 
   exec {'remove the (un)installer':
     command => '/bin/rm -rf /Applications/Install.app /Applications/Uninstall.app',
-    require => [
-                Package['DroboDashboard'],
-                File['/Applications/Install.app'],
-                File['/Applications/Uninstall.app'],
-                File['/Applications/Drobo\ Dashboard.app']
-                ]
+    require => Package['DroboDashboard'],
+    onlyif  => 'test -d /Applications/Install.app && test -d /Applications/Uninstall.app && test -d /Applications/Drobo\ Dashboard.app'
   }
 }
